feat(ui): worklog toolbar + time entry — icon Add, Bulk-entry modal, time fixes
- Add Entry is now an icon-only button (+) with aria-label/title.
- Bulk Entry is reachable from the Worklog toolbar via a modal: extract the
  Extras bulk-entry form into a shared <BulkEntryForm> (Extras page now just
  hosts it), opened in an Ark dialog so you don't leave the grid; on success it
  refetches the entries.
- Editing an empty start cell prefills the suggested start (latest entry's end,
  else now) so the editor opens on a sensible value.
- A terse start/end time is normalized in the cell on commit (1300/13 → 13:00)
  instead of only on the next refetch.
